ROUTINE
ZDVX

+++ Azimutal/Azimutal.h
@@ -46,14 +46,13 @@ private:
 	//constantes de calibração
 	float minPWM, maxPWM;				//min e max leitura analogica do pwm.
 	float restrition;				//no modo normal é a restrição do quanto pode ir ao lado
-	int	  driverConf
-										// float ctrlConf = 100;
+	int	  driverConf									// float ctrlConf = 100;
 
 	//metodos privados
 	float map_f(float number, float minI, float maxI, float minF, float maxF);
 	float readPWM(int pin);
 	bool filter(float x);
-	float readStep(void);
+	int readStep(void);
 
 	float nullHole;
 	const float un = 10;
